requirements.txt
# Core Langchain + Azure OpenAI + Qdrant
langchain-text-splitters
langchain-community
langchain-qdrant
langchain-openai
langgraph
qdrant-client
aiofiles
PyPDF2
python-dotenv
httpx
typing_extensions
asyncio
langchain-pinecone
streamlit



openevals

